AI_agent_tutorial/requirements.txt
langchain
wikipedia
langchain-community
langchain-openai
langchain-anthropic
python-dotenv
pydantic
langchain_google_genai
langchain_core
duckduckgo-search